Finished Projects page, added and finished Contacts page
Finished the projects page, making the GitHub integration fully functional as well as adding new projects to it. Added and finished a small contacts section, with 3 ways to contact me as well as working links. Also added a particle background to the rest of the site and cleaned up a few things.